# docs-site/docs/api/openapi/core-full.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AstroEngine API",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/interpret/rulepacks": {
      "get": {
        "tags": [
          "interpret"
        ],
        "summary": "List uploaded rulepacks.",
        "operationId": "listRulepacks",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-indexed page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-indexed page number."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items to include in the response page (max 100).",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Number of items to include in the response page (max 100)."
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulepackCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "interpret"
        ],
        "summary": "Upload Rulepack",
        "operationId": "uploadRulepack",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulepackMeta"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/interpret/rulepacks/{rulepack_id}": {
      "get": {
        "tags": [
          "interpret"
        ],
        "summary": "Get Rulepack",
        "operationId": "getRulepack",
        "parameters": [
          {
            "name": "rulepack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rulepack Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulepackVersionPayload"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "interpret"
        ],
        "summary": "Delete Rulepack",
        "operationId": "deleteRulepack",
        "parameters": [
          {
            "name": "rulepack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rulepack Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/interpret/rulepacks/lint": {
      "post": {
        "tags": [
          "interpret"
        ],
        "summary": "Lint Rulepack Endpoint",
        "operationId": "lintRulepack",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulepackLintResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/interpret/relationship": {
      "post": {
        "tags": [
          "interpret"
        ],
        "summary": "Run a relationship interpretation.",
        "operationId": "interpretRelationship",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretRequest"
              },
              "examples": {
                "synastry": {
                  "summary": "Synastry findings from pre-computed hits",
                  "value": {
                    "rulepack_id": "relationship_basic",
                    "scope": "synastry",
                    "synastry": {
                      "hits": [
                        {
                          "a": "Sun",
                          "b": "Moon",
                          "aspect": "trine",
                          "severity": 0.6
                        },
                        {
                          "a": "Venus",
                          "b": "Mars",
                          "aspect": "conjunction",
                          "severity": 0.5
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterpretResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CompositePositionsInput": {
        "properties": {
          "positions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Positions"
          }
        },
        "type": "object",
        "required": [
          "positions"
        ],
        "title": "CompositePositionsInput"
      },
      "DavisonPositionsInput": {
        "properties": {
          "positions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Positions"
          }
        },
        "type": "object",
        "required": [
          "positions"
        ],
        "title": "DavisonPositionsInput"
      },
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine readable error code."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human friendly summary of the error."
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Optional structured details that expand on the error."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorEnvelope",
        "description": "Standardized error payload returned by the public API."
      },
...
      "FindingsFilters": {
        "properties": {
          "profile": {
            "type": "string",
            "title": "Profile",
            "default": "balanced"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K",
            "default": 50
          },
          "min_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Min Score",
            "default": 0.0
          },
          "include_tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Tags"
          },
          "exclude_tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclude Tags"
          }
        },
        "type": "object",
        "title": "FindingsFilters",
        "description": "Filtering controls for interpretation results."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterpretRequest": {
        "properties": {
          "rulepack_id": {
            "type": "string",
            "title": "Rulepack Id"
          },
          "scope": {
            "type": "string",
            "enum": [
              "synastry",
              "composite",
              "davison"
            ],
            "title": "Scope",
            "default": "synastry"
          },
          "filters": {
            "$ref": "#/components/schemas/FindingsFilters"
          },
          "synastry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SynastryHitsInput"
              },
              {
                "$ref": "#/components/schemas/SynastryPositionsInput"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synastry"
          },
          "composite": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompositePositionsInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "davison": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DavisonPositionsInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "rulepack_id"
        ],
        "title": "InterpretRequest",
        "description": "Request payload for `/relationship` evaluations."
      },
      "InterpretResponse": {
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "totals": {
            "additionalProperties": true,
            "type": "object",
            "title": "Totals"
          },
          "rulepack": {
            "$ref": "#/components/schemas/RulepackMeta"
          }
        },
        "type": "object",
        "required": [
          "findings",
          "totals",
          "rulepack"
        ],
        "title": "InterpretResponse",
        "description": "Response payload describing evaluated findings."
      },
...
      "ProfileDefinition": {
        "properties": {
          "base_multiplier": {
            "type": "number",
            "title": "Base Multiplier",
            "default": 1.0
          },
          "tag_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Tag Weights"
          },
          "rule_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Rule Weights"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ProfileDefinition",
        "description": "Profile weighting configuration for a rulepack."
      },
...
      "RuleDefinition": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "scope": {
            "type": "string",
            "enum": [
              "synastry",
              "composite",
              "davison"
            ],
            "title": "Scope"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "default": 1.0
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "when": {
            "$ref": "#/components/schemas/RuleCondition"
          },
          "then": {
            "$ref": "#/components/schemas/RuleThen"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scope",
          "then"
        ],
        "title": "RuleDefinition",
        "description": "Single interpretation rule supporting legacy and modern payloads."
      },
...
      "RulepackCollection": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RulepackMeta"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "RulepackCollection",
        "description": "Paginated rulepack listing.",
        "example": {
          "items": [
            {
              "available_versions": [
                1
              ],
              "created_at": "2023-01-01T00:00:00+00:00",
              "description": "Core synastry findings",
              "id": "relationship_basic",
              "mutable": false,
              "name": "Relationship Basic",
              "title": "Relationship Essentials",
              "version": 1
            }
          ],
          "page": 1,
          "page_size": 25,
          "total": 1
        }
      },
      "RulepackLintResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "errors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Errors"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "RulepackLintResult",
        "description": "Lint outcome for a rulepack upload attempt."
      },
      "RulepackMeta": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "mutable": {
            "type": "boolean",
            "title": "Mutable",
            "default": false
          },
          "available_versions": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Available Versions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "title",
          "created_at"
        ],
        "title": "RulepackMeta",
        "description": "Metadata describing a stored rulepack version."
      },
      "RulepackVersionPayload": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/RulepackMeta"
          },
          "profiles": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ProfileDefinition"
            },
            "type": "object",
            "title": "Profiles"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/RuleDefinition"
            },
            "type": "array",
            "title": "Rules"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "etag": {
            "type": "string",
            "title": "Etag"
          },
          "content": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content"
          },
          "mutable": {
            "type": "boolean",
            "title": "Mutable"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "profiles",
          "rules",
          "version",
          "etag",
          "content",
          "mutable"
        ],
        "title": "RulepackVersionPayload",
        "description": "Envelope returned when fetching a rulepack."
      },
...
      "SynastryHitsInput": {
        "properties": {
          "hits": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Hits"
          }
        },
        "type": "object",
        "required": [
          "hits"
        ],
        "title": "SynastryHitsInput",
        "description": "Direct synastry hits payload."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}